[UR] Remove UMF sources and use standalone UMF repo instead
UMF has become a separate project with its own repository.
It is currently under active development and is intended to
be used by various oneAPI components, not just UR. All
further enhancements and fixes will only go into the UMF
repository - we do not plan to merge such changes here in
the UR repository and into the UMF repository separately.