[CTS] Added test for sequencing events
This tests that the `urEnqueueEventsWaitWithBarrier` function
functions as a barrier. That is, it blocks until all events prior
to it in the queue are completed. Since this test may fail
non-deterministically, it is ran a few times.

Note that current adapters seem to always execute events
sequentially, so this is a no-op.