[L0 v2] implement support for images
Use dynamic_cast to convert ur_mem_handle_t to
buffer or image. This is not ideal, but buffer and image
are pretty much separate entites that just happen to
share the same handle type. There is no shared impl
between buffer and image, expcept for a few lines in
urMemGetInfo.

Alternative to dynamic_cast and inheritance would be
to just use std::variant, but this is problematic since
ur_mem_buffer_t has subclasses and relies on polymorphic
behavior. Variant would have to hold pointers to buffer/image
adding one more indirection.